Answers:
You can now list answers, add answers, and attach / remove answers to questions. More work on the basic UI to make it faster to add multiple new answers, auto-add tags to answers, filter the list of available answers by tag, and so much more, but the CRU-D is coming along.